IdentityService was added

+++ src/AuctionService/Controllers/AuctionController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -52,12 +53,13 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto createAuctionDto)
         {
             var auction = _mapper.Map<Auction>(createAuctionDto);
-            // TODO: add current user as a seller
-            auction.Seller = "test";
+            
+            auction.Seller = User.Identity.Name;
 
             _context.Auctions.Add(auction);
 
@@ -73,6 +75,7 @@ namespace AuctionService.Controllers
                 new { auction.Id }, newAuction);
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
@@ -83,7 +86,8 @@ namespace AuctionService.Controllers
             if (auction == null)
                 return NotFound();
 
-            // TODO: check seller == username
+            if(auction.Seller != User.Identity.Name)
+                return Forbid();
 
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -101,6 +105,7 @@ namespace AuctionService.Controllers
             return BadRequest("Failed to update auction");
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
@@ -110,7 +115,9 @@ namespace AuctionService.Controllers
             if (auction == null)
                 return NotFound();
 
-            // TODO: check seller == username
+            if(auction.Seller != User.Identity.Name)
+                return Forbid();
+
             _context.Auctions.Remove(auction);
 
             await _publishEndpoint.Publish<AuctionDeleted>(new { Id = id.ToString() });
